==> 01_Data Collection for DEPRESSION.ipynb <==
# Data Collection

import requests
import time
import pandas as pd
import json

## Starting with Depression Subreddit

url = 'https://www.reddit.com/r/depression/.json'

headers = {'User-agent': 'VIP'}

res = requests.get(url, headers = headers)
res.status_code

#this will return a dictionary 
the_json = res.json()

## Keys of first url (depression subreddit)

sorted(the_json.keys())

sorted(the_json['data'].keys())

## Data of first post

the_json['data']['children'][0]
#my first child (aka post)

## All usernames for this scrape

[post['data']['name'] for post in the_json['data']['children']]
#shows all usernames 

len([post['data']['name'] for post in the_json['data']['children']])
#27 blogs per scrape

#so the last usernmae needs to tacked on after the .json? url 
#www.reddit.com/hot.json?after=THE_AFTER_FROM_ABOVE
#key= value pattern

#this way will help me get more blog posts 

## Loop to get ~1000 unique blog post data

param = {'after': 'last_username'}

requests.get(url, params = param, headers = headers)
#this get another page of results 



post = []
after = None
for i in range(100): #hitting it 100 times 
    print(i)
    if after == None:
        params = {}
    else:
        params = {'after': after}
    url = 'https://www.reddit.com/r/depression/.json'
    res = requests.get(url, params = params, headers = headers)
    if res.status_code == 200:
        the_json = res.json()
        post.extend(the_json['data']['children'])
        after = the_json['data']['after']
    else: 
        print(res.status_code)
        break  
    
    time.sleep(0.5) 

df = pd.DataFrame(post)

df.head()

df['data'][0].keys()

df['data'][0]['subreddit']

len(df['data'])

data = df['data'].apply(pd.Series)

data.shape

data.head()

data['selftext'].isnull().sum()

data.columns

data.to_csv('./datasets/depression.csv')
