==> 02_Does our subreddit repost blogs?.ipynb <==
# Objective: Each subreddit must have 1000 unique selftexts
---

import pandas as pd

# Depression Data

depression = pd.read_csv('./datasets/depression.csv')

depression.head()

## First Size of Depression Dataframe

depression.shape

## Length of Dataframe without duplicates

#i want to drop all selftext duplicates since this subreddit reposts blogs
unique = depression.drop_duplicates(subset = 'selftext')

#length of unique selftexts
len(unique)

unique.head()

### Since we have 923 unique self texts, lets go back to our Depression Data Collection notebook and get more blog posts in order to reach 1000 unique self texts 

## Trying Anxiety Subreddit


anxiety = pd.read_csv('./datasets/anxiety.csv')
anxiety.shape

#this is the length of unique selftexts from anxiety
unique = anxiety.drop_duplicates(subset = 'selftext')
len(unique)

#### We have a decent amount of unique blogs for the Anxiety Subreddit! Let's proceed with Anxiety!

# Conclusion

**The two subreddits we'll use for this project are: r/Anxiety and r/Depression**
- Let's export both dataframes with unique selftexts as csv files
- The four columns I want to keep are: subreddit, selftext, title, created_utc

### Anxiety 

anxiety.shape

unique_anxiety = anxiety.drop_duplicates(subset = 'selftext')

unique_anxiety.shape

type(unique_anxiety)

#creating unique anxiety df with selftext, title, and created utc
features = ['subreddit', 'selftext', 'title', 'created_utc']
unique_anxiety = unique_anxiety[features]

unique_anxiety.head()

unique_anxiety.shape

#creating unique anxiety df
unique_anxiety.to_csv('./datasets/unique_anxiety.csv')

### Depression

unique_depression = depression.drop_duplicates(subset = 'selftext')

unique_depression.shape

#creating unique df for derpression 
features = ['subreddit', 'selftext', 'title', 'created_utc']
unique_depression = unique_depression[features]

unique_depression.head()

unique_depression.shape

#creating unique depression df
unique_depression.to_csv('./datasets/unique_depression.csv')
