==> 03_Merging Dataframes.ipynb <==
# Objective
- Merge both dataframes and change subreddit column to binary since it's our target variable 

import pandas as pd

#reading anxiety
anxiety = pd.read_csv('./datasets/unique_anxiety.csv')

#reading depression
depression = pd.read_csv('./datasets/unique_depression.csv')

### Let's merge selftext and title for Anxiety

anxiety.shape

anxiety.head()

#lets gather selftext and title since we're only analyzing words 
anxiety['text'] = anxiety['selftext'] +  ' ' + anxiety['title']

anxiety['text'][0]

### Let's merge selftext and title for Depression

depression.shape

#lets gather selftext and title since we're only analyzing words 
depression['text'] = depression['selftext'] +  ' ' + depression['title']

depression['text'][0]

depression.head()

### For each dataframe, let's get `subreddit` and `text` in order to merge depression and anxiety

#depression
depression_features = ['subreddit', 'text']
depression_2 = depression[depression_features]
depression_2.head()

depression_2.shape

#anxiety
anxiety_features = ['subreddit', 'text']
anxiety_2 = anxiety[anxiety_features]
anxiety_2.head()

anxiety_2.shape

## Time to merge both data frames

frames = [depression_2, anxiety_2]

df = pd.concat(frames)

df.shape

df.head()

df.isnull().sum()

df.dropna(axis = 0, inplace = True)

df.shape

df.isnull().sum()

df['subreddit'].value_counts()

## Our df has non null values and contains both subreddits
- Let's make depression = 1 and anxiety = 0
- We'll export it as a csv called df.csv

df.head()

#changing sbreddit column to binary variable 
df['subreddit'] = df['subreddit'].map({'depression': 1, 'Anxiety': 0})

df.head()

df['subreddit'].value_counts()

df.to_csv('./datasets/df.csv')
