==> 04_Data Text Cleaning and PreProcessingTHE_ONE.ipynb <==
# Data Text Cleaning 
In this notebook, we will execute the following: 
- Remove HTML artifacts
- Remove punctuation 
- Lowercase everything 
- Tokenize then Snowball all texts 
- Save cleaned df as csv

import nltk

from nltk.stem import SnowballStemmer
from nltk.tokenize import RegexpTokenizer
from bs4 import BeautifulSoup

import pandas as pd

df = pd.read_csv('./datasets/df.csv')

df.shape

df.head()

### Removing HTML artifacts 

example1 = BeautifulSoup(df['text'][0])
example1.get_text()

#lets use beautiful soup to remove html language
df['text'] = [BeautifulSoup(text).get_text() for text in df['text'] ]

df.head()

## Tokenizing and Snowballing

#### Let's Tokenize First

#instntiate tokenizer
tokenizer = RegexpTokenizer(r'\w+') 

#instntiate snowballstemmer
snow = SnowballStemmer(language = 'english')

#shoutout to stack overflow 
#this code will tokenize then stem the lower case of text
def snow_text(text):
    return [snow.stem(w.lower()) for w in tokenizer.tokenize(text.replace("'", ""))]  #add .replace("'", "")


df['snow_text'] = df['text'].apply(snow_text)

df.head()

df['text'][2].replace("'", "")

### Checking if it got Snowballed

df['text'][2]

df['snow_text'][2]

df.shape

df.head()

#lets join each word in snow_text by spaces 
df['texty'] = [[' '.join(i)] for i in df['snow_text']]

df.head(1)

#lets convert to string 
df['string'] = [str(i) for i in df['texty']]

df.head(1)

#removing left bracket, right bracket, and apostrophe
df['stringOG'] = [i.replace('[', '').replace(']', '').replace("'", '') for i in df['string']]

df.head()

## Saving Snow Data 
- Called snow.csv

#remember i am only analyzing text, therefore iwant to keep the string and subreddit
features = ['subreddit', 'stringOG']
dataframe = df[features]

dataframe.head()

dataframe.shape

#saving it to df
dataframe.to_csv('./datasets/snow.csv')
