==> 05_Vader StepTHE_ONE.ipynb <==
# Objective
- Add vader sentiment analysis 
- TfidfVectorizer

import nltk

from nltk.sentiment import vader

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression

df = pd.read_csv('./datasets/snow.csv')
df.head()

df.shape

df['subreddit'].value_counts()

# Let's Add Vader Sentiment Analysis 

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

#instantiate vader
analyser = SentimentIntensityAnalyzer()

#getting polarity scores(4) from stringOG
df['vader'] = [analyser.polarity_scores(text) for text in df['stringOG']]

df.head()

df['vader'][0]

#this code creates new columns of vader scores 
df['neg'] = df['vader'].apply(lambda x: x.get('neg'))
df['neu'] = df['vader'].apply(lambda x: x.get('neu'))
df['pos'] = df['vader'].apply(lambda x: x.get('pos'))
df['compound'] = df['vader'].apply(lambda x: x.get('compound'))

df.head()

#lets save the df with vader columns
df.to_csv('./datasets/df_vader.csv')

# Train Test Split

#first stage of train test split 
features = ['stringOG', 'neg', 'neu', 'pos', 'compound']
X = df[features]
y = df['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, stratify = y)

X_train

# Tfidf String

#separating string from vader score
#im doing this because I need to Tfidfvectorize stringOG then append the vader scores
X_train_string = X_train['stringOG']
X_test_string = X_test['stringOG']

tfidf = TfidfVectorizer()
tfidf.fit(X_train_string)

# transofrming 
train_transform = tfidf.transform(X_train_string)
test_transform = tfidf.transform(X_test_string)

#creating df of transformed data
train_transform_df = pd.DataFrame(train_transform.toarray(),
                   columns= tfidf.get_feature_names())

test_transform_df = pd.DataFrame(test_transform.toarray(),
                   columns= tfidf.get_feature_names())

train_transform_df.shape

train_transform_df.head()

#appending to train data
vader = ['neg', 'neu', 'pos', 'compound']
X_train_df = pd.concat(objs = [train_transform_df, X_train.reset_index()[vader]], axis = 1)

X_train_df.isnull().sum().sum()

X_train_df.head()

#appending test data 
X_test_df = pd.concat(objs = [test_transform_df, X_test.reset_index()[vader]], axis = 1)

X_test_df.head()

X_test_df.isnull().sum().sum()

## Append Target Variable for both train and test data

#adding subreddit to train 
X_train_df['subreddit'] = y_train.reset_index()['subreddit']

X_train_df.head()

X_train_df.isnull().sum().sum()

#adding subreddit to test
X_test_df['subreddit'] = y_test.reset_index()['subreddit']

X_test_df.head()

X_test_df.isnull().sum().sum()

# Lets save it to csv file

X_train_df.to_csv('./datasets/train_snow.csv')

X_test_df.to_csv('./datasets/test_snow.csv')
