==> 06_Modeling & Metrics.ipynb <==
# Modeling
These are the three following models used to classify subreddits:
- Logistic Regression 
- Naive Bayes Classification(2)
- Random Forest 
- GradientBooster


## Starting with Logistic Regression 

import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV, cross_val_score
from sklearn.metrics import confusion_matrix, f1_score
from sklearn.ensemble import GradientBoostingClassifier

train = pd.read_csv('./datasets/train_snow.csv')
train.head()

test = pd.read_csv('./datasets/test_snow.csv')
test.head()

train.shape

test.shape

#dropping unnamed columns
train.drop('Unnamed: 0', axis = 1, inplace= True)
test.drop('Unnamed: 0', axis = 1, inplace= True)

train.head()

test.head()

## Model Prep

#features are everything except for subreddit
features = [col for col in train.columns if col!= 'subreddit']
X_train = train[features]

#features are everything except for subreddit
features = [col for col in test.columns if col!= 'subreddit']
X_test = test[features]

#y variable is subreddit
y_train = train['subreddit']

#y variable is subreddit
y_test = test['subreddit']

## Instantiate Model

logreg = LogisticRegression()

logreg.fit(X_train, y_train)

logreg.score(X_train, y_train)

logreg.score(X_test, y_test)

#### Confusion Matrix for Logistic Regression Model

pred = logreg.predict(X_test)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score 

f1_score(y_test, pred)

## Naive Bayes Time
- Under Naive Bayes, we will use 2 models: Gaussian and Bernoulli

from sklearn.naive_bayes import GaussianNB, BernoulliNB

### Gaussian Naive Bayes & Bernoulli

g = GaussianNB()
b = BernoulliNB()

cross_val_score(g, X_train, y_train, cv = 5).mean()

cross_val_score(b, X_train, y_train, cv = 5).mean()

## Gaussian

g.fit(X_train, y_train)

g.score(X_train, y_train)

g.score(X_test, y_test)

#### Confusion Matrix for Gaussian Model

pred = g.predict(X_test)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)

## Bernoulli

b.fit(X_train, y_train)

b.score(X_train, y_train)

b.score(X_test, y_test)

#### Confusion Matrix for Bernoulli Model

pred = b.predict(X_test)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)

## Random Forest 

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier()
rf.fit(X_train, y_train)

rf.score(X_train, y_train)

rf.score(X_test, y_test)

#### Confusion Matrix for Random Forest 

pred = rf.predict(X_test)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)

## Gradient Booster

gb = GradientBoostingClassifier()
gb.fit(X_train, y_train)

gb.score(X_train, y_train)

gb.score(X_test, y_test)

#### Confusion Matrix for Gradient Booster 

pred = gb.predict(X_test)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)
