==> 07_Voting ClassifierTHE_ONE.ipynb <==
## Voting Classifier
- Logistic
- Random Forest
- Gradient Booster


import pandas as pd
from sklearn.ensemble import GradientBoostingClassifier, AdaBoostClassifier, VotingClassifier
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier

train = pd.read_csv('./datasets/train_snow.csv')
train.head()

test = pd.read_csv('./datasets/test_snow.csv')
test.head()

train.drop('Unnamed: 0', axis = 1, inplace = True)
test.drop('Unnamed: 0', axis = 1, inplace = True)

train.head()

test.head()

features = [col for col in train.columns if col!= 'subreddit']
X_train = train[features]
y_train = train['subreddit']

features = [col for col in test.columns if col!= 'subreddit']
X_test = test[features]
y_test = test['subreddit']

vote = VotingClassifier([
    ('rf', RandomForestClassifier(max_features = None)),
    ('logreg', LogisticRegression(C = 1, penalty = 'l2')),
    ('grad', GradientBoostingClassifier())  #main ones (n_estimators, learning rate, max_depth)
])

vote_param = {
   
    'rf__n_estimators': [100, 120],
    'rf__max_depth': [4,5],
    'grad__n_estimators': [250, 300],
    'grad__max_depth': [3, 4]
}

gs = GridSearchCV(vote, param_grid = vote_param , cv=5)
gs.fit(X_train, y_train)
print(gs.best_score_)
gs.best_params_

gs.score(X_train, y_train)

gs.score(X_test, y_test)

from sklearn.metrics import confusion_matrix, f1_score

pred = gs.predict(X_test)
cm = confusion_matrix(y_test, pred)

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

f1_score(y_test, pred)

dictionary = dict(zip(tfidf.get_feature_names(), model.coef_))

gs.best_estimator_  #will tell me the best hyperparameter
