==> 09_TRANSFORMCLASS.ipynb <==
# Modeling by GridSearching
These are the three following models used to classify subreddits:
- Logistic Regression 
- Naive Bayes Classification(2)
- Random Forest 
- GradientBooster

# Objective
- Gridsearch to find optimal parameters to ultimately execute VoteClassifier using the top 3 models

## Starting with Logistic Regression 

import nltk

import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV, cross_val_score, train_test_split
from sklearn.metrics import confusion_matrix, f1_score
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline

df = pd.read_csv('./datasets/df_vader.csv')
df.head()

df.shape

df.drop(['Unnamed: 0', 'Unnamed: 0.1'], axis = 1, inplace= True)

df.head()

## Train Test Split

features = ['stringOG', 'vader']
X = df[features]
y = df['subreddit']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=42, stratify =y)

X_train_p1 = X_train['stringOG']
X_test_p1 = X_test['stringOG']

## Creating Class to append vader columns before modeling after tfidf

from sklearn.base import TransformerMixin
from sklearn.pipeline import Pipeline, FeatureUnion

class Appender(TransformerMixin):
    def __init__(self, *_):
        pass
   
        
    def transform(self, X, *_):
        X = pd.DataFrame(X.toarray())
        #if score is train:
        X['neg'] = X_train.reset_index()['vader'].apply(lambda x: eval(x).get('neg'))  #eval takes string and evaluates as python code
        X['neu'] = X_train.reset_index()['vader'].apply(lambda x: eval(x).get('neu'))
        X['pos'] = X_train.reset_index()['vader'].apply(lambda x: eval(x).get('pos'))
        X['compound'] = X_train.reset_index()['vader'].apply(lambda x: eval(x).get('compound'))
        #else: use X_test vader 
        return X
    
    def fit(self, *_):
        return self
    

## Let's Try to Pipeline our GridSearch to Get a Better Score

## Starting with Logistic

pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('append', Appender()),
    ('lr', LogisticRegression())
])



pipe_params = {
    'lr__penalty': ['l1', 'l2'],
    'lr__C': [0.2, 0.6, 1.0]
    #can add C parameter for logisitic regression
}


gs = GridSearchCV(pipe, param_grid=pipe_params, cv=3)


gs.fit(X_train_p1, y_train)


print(gs.best_score_)
gs.best_params_

gs.score(X_train_p1, y_train)

gs.score(X_test_p1, y_test)

#how to compensate for test vader...isnt it using the vader score for train? 
#what happens to test vader?

tfidf = TfidfVectorizer()

tfidf.fit(X_train_p1, y_train)

train = tfidf.transform(X_train_p1)
test = tfidf.transform(X_test_p1)

train_df = pd.DataFrame(train.toarray(), columns= tfidf.get_feature_names())
test_df = pd.DataFrame(test.toarray(), columns= tfidf.get_feature_names())

X_train['neg'] =  X_train['vader'].apply(lambda x: eval(x).get('neg'))
X_train['neu'] =  X_train['vader'].apply(lambda x: eval(x).get('neu'))
X_train['pos'] =  X_train['vader'].apply(lambda x: eval(x).get('pos'))
X_train['compound'] =  X_train['vader'].apply(lambda x: eval(x).get('compound'))

features = ['neg', 'neu', 'pos', 'compound']
X_train_vader = X_train[features]

X_test['neg'] =  X_test['vader'].apply(lambda x: eval(x).get('neg'))
X_test['neu'] =  X_test['vader'].apply(lambda x: eval(x).get('neu'))
X_test['pos'] =  X_test['vader'].apply(lambda x: eval(x).get('pos'))
X_test['compound'] =  X_test['vader'].apply(lambda x: eval(x).get('compound'))

features = ['neg', 'neu', 'pos', 'compound']
X_test_vader = X_test[features]

X_train_df = pd.concat(objs=[train_df, X_train_vader.reset_index()], axis = 1).drop('index', axis = 1)

X_train_df.head()

X_test_df = pd.concat(objs=[test_df, X_test_vader.reset_index()], axis = 1).drop('index', axis = 1)

#### Logistic Regression 

logreg = LogisticRegression(C = 1.0, penalty = 'l2')
logreg.fit(X_train_df, y_train)

logreg.score(X_train_df, y_train)

logreg.score(X_test_df, y_test)

#### Confusion Matrix for Logistic Regression Model

pred = logreg.predict(X_test_df)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score 

f1_score(y_test, pred)

## Random Forest 

from sklearn.ensemble import RandomForestClassifier

pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('append', Appender()),
    ('rf', RandomForestClassifier())
])

pipe_params = {
    'rf__max_features': [None, 'log2', 'auto'],
    'rf__max_depth': [3, 4, 5],
    'rf__n_estimators': [100, 200, 300]
    
}


gs = GridSearchCV(pipe, param_grid=pipe_params, cv=3)


gs.fit(X_train_p1, y_train)


print(gs.best_score_)
gs.best_params_

gs.score(X_train_p1, y_train)

gs.score(X_test_p1, y_test)

tfidf = TfidfVectorizer()

tfidf.fit(X_train_p1, y_train)

train = tfidf.transform(X_train_p1)
test = tfidf.transform(X_test_p1)

train_df = pd.DataFrame(train.toarray(), columns= tfidf.get_feature_names())
test_df = pd.DataFrame(test.toarray(), columns= tfidf.get_feature_names())

X_train_df = pd.concat(objs=[train_df, X_train_vader.reset_index()], axis = 1).drop('index', axis = 1)

X_test_df = pd.concat(objs=[test_df, X_test_vader.reset_index()], axis = 1).drop('index', axis = 1)

rf = RandomForestClassifier(max_depth= 5, max_features= None, n_estimators= 100)
rf.fit(X_train_df, y_train)

rf.score(X_train_df, y_train)

rf.score(X_test_df, y_test)

#### Confusion Matrix for Random Forest 

pred = rf.predict(X_test_df)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)

## Gradient Booster

from sklearn.ensemble import GradientBoostingClassifier

pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('append', Appender()),
    ('gb', GradientBoostingClassifier())
])

pipe_params = {
    'gb__learning_rate': [0.05, 0.1, 0.15],
    'gb__max_depth': [3, 4, 5],
    'gb__n_estimators': [100, 200, 300]
    
}


gs = GridSearchCV(pipe, param_grid=pipe_params, cv=3)


gs.fit(X_train_p1, y_train)


print(gs.best_score_)
gs.best_params_

gs.score(X_train_p1, y_train)

gs.score(X_test_p1, y_test)

tfidf = TfidfVectorizer()

tfidf.fit(X_train_p1, y_train)

train = tfidf.transform(X_train_p1)
test = tfidf.transform(X_test_p1)

train_df = pd.DataFrame(train.toarray(), columns= tfidf.get_feature_names())
test_df = pd.DataFrame(test.toarray(), columns= tfidf.get_feature_names())

X_train_df = pd.concat(objs=[train_df, X_train_vader.reset_index()], axis = 1).drop('index', axis = 1)

X_test_df = pd.concat(objs=[test_df, X_test_vader.reset_index()], axis = 1).drop('index', axis = 1)

gb = GradientBoostingClassifier(learning_rate= 0.1, max_depth= 3, n_estimators= 300)
gb.fit(X_train_df, y_train)

gb.score(X_train_df, y_train)

gb.score(X_test_df, y_test)

#### Confusion Matrix for Gradient Booster 

pred = gb.predict(X_test_df)
cm = confusion_matrix(y_test, pred)

#### Confusion Dataframe

cm_df = pd.DataFrame(cm, 
             columns = ['predicted depression', 'predicted anxiety'], 
             index = ['actual depression', 'actual anxiety'])

cm_df

#### F1 Score

f1_score(y_test, pred)
